+++ cli.ts
@@ -14,7 +14,7 @@ import {
 } from './cli/mod.ts'
 import 'https://deno.land/std@0.179.0/dotenv/load.ts'
 
-export const version = 'v0.4.3'
+export const version = 'v0.4.4'
 
 const command = new Command()
 	.name('arkiver')